Add config and controller component
