+++ Shared/config.json
@@ -1,0 +1,25 @@
+{
+    "oled": {
+        "height": 64,
+        "width": 128
+    },
+    "urlThingSpeak": "https://api.thingspeak.com/update?api_key=LCSJ5M98PSD7T1YT",
+    "urlFirebase": "https://agro-app-2aa55-default-rtdb.firebaseio.com/",
+    "wifiConnect": {
+        "name": "TP-Link_D54E",
+        "pass": "04577041"   
+    },
+    "pines": {
+        "dth": 4,
+        "bomb": 5,
+        "ft": 36,
+        "ldr": 39,
+        "screen": {
+            "scl": 22,
+            "sda": 21
+        },
+        "ledBomb": 2,
+        "ledHum": 4
+    }
+
+}
